+++ axual-webclient/applications.go
@@ -52,15 +52,17 @@ func (c *Client) CreateApplication(data ApplicationRequest) (*ApplicationRespons
 	return &o, nil
 }
 
+func (c *Client) GetApplicationByNameOrShortName(params url.Values) (*ApplicationResponse, error) {
+	o := ApplicationResponse{}
+	suffix := fmt.Sprintf("findByName?name=%s", params.Get("name"))
+	if params.Get("shortName") != "" {
+		suffix = fmt.Sprintf("findByShortName?shortName=%s", params.Get("shortName"))
+	}
+	endpoint := fmt.Sprintf("%s/applications/search/%s", c.ApiURL, suffix)
 
-func (c *Client) GetApplicationsByAttributes(attributes url.Values) (*ApplicationsByAttributesResponse, error) {
-    o := ApplicationsByAttributesResponse{}
-
-    url := fmt.Sprintf("%s/applications/search/findByAttributes?%s", c.ApiURL, attributes.Encode())
-	fmt.Println("URL", url)
-    err := c.RequestAndMap("GET", url, nil, nil, &o)
-    if err != nil {
-        return nil, err
-    }
-    return &o, nil
+	err := c.RequestAndMap("GET", endpoint, nil, nil, &o)
+	if err != nil {
+		return nil, err
+	}
+	return &o, nil
 }