+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 * Fixed `axual_topic_config` to allow in-place updates of `key_schema_version` and `value_schema_version` fields instead of forcing resource replacement
 * Added boolean `force` attribute to `axual_topic_config` resource to force updates with incompatible schema version changes
 * Retrieve an application using `findByName` or `findByShortName` endpoints instead of `findByAttributes`
+* Fix state inconsistency bugs in `axual_application_access_grant` and `axual_application_deployment` resources
 
 ## [2.5.6](https://github.com/Axual/terraform-provider-axual/releases/tag/v2.5.6) - 2025-06-06
 * Add User data source